Conflicts:
	_backup/data/postprocessing/background.frag
	_backup/gloperate-text/source/FontImporter.cpp
	_backup/gloperate/include/gloperate/stages/ColorGradientSelectionStage.h
	_backup/gloperate/include/gloperate/tools/ColorGradientPreparation.h
	_backup/gloperate/source/base/ExplicitEquidistantColorGradient.cpp
	_backup/gloperate/source/stages/ColorGradientSelectionStage.cpp
	_backup/gloperate/source/tools/ColorGradientPreparation.cpp
	data/gloperate/qml/Viewer.qml
	source/applications/gloperate-viewer/main.cpp
	source/gloperate/include/gloperate/viewer/RenderSurface.h
	source/gloperate/source/viewer/RenderSurface.cpp

+++ source/gloperate-ffmpeg/source/VideoTool.cpp
@@ -64,7 +64,7 @@ void VideoTool::createVideo(std::function<void(int, int)> progress, bool glConte
 {
     auto deviceViewport = m_surface->deviceViewport();
     auto virtualViewport = m_surface->virtualViewport();
-    auto vp = glm::ivec4(0, 0, m_width, m_height);
+    auto vp = glm::vec4(0, 0, m_width, m_height);
     auto length = m_length * m_fps;
 
     m_surface->onViewport(vp, vp);
